Set correct action for periodic ResourceUpdate task
In the past action was set to PRIORITY_SYNC_ROUTERS_TASK, which
has the value of 2, which again maps to the action DELETE_RELATED_ROUTER,
which is not what we want to do here. Instead we now set it to
ADD_UPDATE_ROUTER. This is also done in the upstream Neutron Queens code
for their agent/l3/agent.py fetch_and_sync_all_routers() task.