Configure TOS based routing loop filter on VRFs
We want to avoid traffic loops, where we forward external traffic
towards the same external interface. In this commit we take care
that internal traffic is marked with precedence "routine" and filters
based on EXT-TOS on the outside BDI. Config already present on the
device will take care of the rest.