Don't stop loopingcalls on exception
If loopincalls in the driver stop on an exception they, well, stop.
Especially for the device_alive_check this means that devices that
once are marked as unreachable will never become reachable again
and the driver stops working alltogether. Adding stop_on_exception=False
to all loopingcalls should stop this behaviour.