Add stack to backtrace of certain exceptions
Exception-logging only contains the backtrace from the point where
an exception is thrown up to where it is catched. We also want to
know the rest of the stack, therefore we append the current stack
to the backtrace. As a tb frame is not modifiable we need to
encapsulate it.