Pass context to all functions doing device-dependant operations
As YANG models sometimes change between ASR firmware versions
the functions generating the XML need to be changed, depending
upon the firmware version running on the device. This might be
a different version, even on the same ASR1k pair. Therefore we
need to pass a context to all these functions so they can make an
informed decision.

Additionally to context being passed this argument was made
mandatory for other functions for it to become clear that it is
actually always necessary to pass some sort of context for a
version check to work.