Implement version check for device context
The version check was previously done on a per connection basis,
but in some cases we need to know the version before a connection
is available for that specific part of the program. Therefore
a connection is now established when someone asks for the version
and it is set directly on the ASR1KContext object. This leaves us
room to later on implement more version dependant branches.