Improve traceback-stacktrace augmentation by allowing details
The exc_info_full() function now accepts parameters to override the
exception type or the exception description. This is usefull when
no exception was thrown, but we still wish to add a stacktrace
to a log message.

It also allows to specify how many frames from the stacktrace should
be skipped.