Use FakeASR1kContext where no device is available, log __eq__
In __str__, __repr__ or __eq__ there is no device context available.
For this staying usable we need to pin down the version that is
used in these cases. For __eq__ this might be dangerous, therefore
we log calls to this function. If it is not used anywhere we can
just remove it.