16.12 compat: Adjst PrefixSeq for new YANG models
We currently need to differentiate in the 16.12 case between two
sets of variables. One set is used by the driver when it creates
this object, the other is gathered from the device itself. When
the 16.12 migration is done we can adjust the way this object
is instanciated by the driver and remove this workaround inside
to_dict().