16.12 compat: Adjust RouteMap/MapSequence for new YANG models
The YANG models for a route map changed in a very weird way. This
commit contains the workarounds for that.

What apparently happened is that <next-hop-addr> was thrown away,
instead there are now two <address> elements inside <next-hop>,
one being the next hop and the other being... the forced flag. To
parse this weirdness, __parameters__() were not enough, therefore
MapSequence now as a custom from_json() method that translates this
into the old 16.9 format when creating this model with data from
a 16.12 device.