Clear RD from VRF when no longer in use
When a RD is reallocated, but still in use by a VRF on the device
we used to delete the VRF. This often does not work out and an
error here stops router provisioning. Therefore we now as a first
step clear the RD from the VRF that is no longer in use. Afterwards
we also try to delete said VRF, but ignore any failure in doing so.