Throw proper exception when versions cannot be determined
Versioning is determined by the YANG capability list, which
requires an established connection to the device. When the driver
starts it often takes more than one attempt to connect to the device,
as an old driver instance is often still hogging the connections.
This results in connection being None, which we now handle properly.