Ensure ml2 sync_loop keeps running by catching BaseException
Catching BaseException is considered bad style and this still holds
true. Unfortunately some of the driver's core exceptions inherit from
BaseException. This caused the ml2 sync_loop thread to die if such an
exception is thrown. Therefore we now catch BaseException instead of
Exception.

The other solution to this would be to let inherit the core exceptions
from Exception, but this would require a minor rework and a reevaluation
of exception handling in general inside the driver, especially in places
where we differentiate between catching Exception and BaseException, so
this task is put off for another time.