BDIInterface now renamed + supports BDI and BD-VIF
This is the first step towards BD-VIF support. As we need parallel
operation of devices that do and do not support BD-VIF we need an
abstraction that is capable of generating config for both interface
types. Therefore BDIInterface is now renamed to VirtualBridgeInterface.
Dependant on the device context either BDI or BD-VIF is used and
configured. This commit contains no migration code towards BD-VIF
and no extra config needed to make BD-VIF actually work.