Add ml2 rpc to get networks with their ports + asr1k info
The ml2 part of this driver will focus more on syncing networks
than only individual ports in the future. For this, the syncloop
needs to iterate over all networks. Therefore this RPC method
is needed to get the networks with all asr1k ports in them plus
their segmentation ids.