Override all bd members when full config is available
In many cases we do not know all bridge members and therefore can
only operate on the members we know that should/shouldn't be present.
When we have the full picture (and has_complete_member_config is
specified) we can remove all other members, therefore issue a
put/replace on the device.