Add extra ServiceInstance type to keep a brdigedomain up
On our current firmware (V1612_1A_SAP) a bridgedomain does not stay
up when there is not at least one "real" interface in that bridge.
As we need our bridgedomain ready to use in case of a failover we
plug another service instance on the always-up Port-channel2 into
that bridgedomain.

This is the YANG model for that endeavour.