Introduce BridgeDomain neutron abstraction for ml2 driver
With 16.12 the ml2 part of the driver needs to know about the
concept of a BridgeDomain on a router and needs to manage its
members. In this commit we use the already present netconf yang
BridgeDomain abstraction to build the Neutron abstraction. The
BridgeDomain class can also handle the old way of binding ports
(with the lb ports / hyperloop cable).