Replace ml2 port based sync with network sync
On 16.12 with BDVif we now do a network based sync instead of a
port based sync in the ml2 driver. This sync knows all the members
of a bridgedomain and can therefore throw out old members.

As we have far less networks than we have ports in a region this leads
to a full sync being much faster, but also more work.