Ignore description for ServiceInstances when unset
When we delete a port in the ml2 driver we need to call update on a
BridgeDomain object to remove the bdvif members from that bridge. At
this point we do not know the network id of the ports that are about
to be deleted. As the network_id is only relevant for the interface
description of the ServiceInstances we can just not update the
description on the device if network_id is not passed along.