Filter out None from bdvif_members when fetching BD from device
When there are no entries on the device for an attribute that has the
type list, NyBase puts a None in a list for this attribute. This happens
in NyBase.from_json(), which constructs the object. As I don't want to
take care of a refactoring in a refactoring in a refactoring while
implementing 16.12 I decided to filter out this value in the
BridgeDomain constructor for now.