Rework cleaning loop
This commit rewrites the asr1k cleaning loop.

Formely the loop pulls down the device config (twice) and then grabs all
entities from it. The call to get the full config takes quite long
(40-80 seconds in some regions) and also contains a lot of information
that is not needed by the driver.

The new cleaning loop fetches each entity type by itself in a single
call and only fetches stubs of the entity, containing everything to
determine the entities' orphan state. If the entity is an orphan a full
copy of the entity will be fetched from the device and it will
subsequently get deleted.