Remove BD membership on BD-VIF deletion
A BD-VIF needs to be removed from the bridge it is a member of, else it
cannot be deleted. Therefore we add this as a postflight for interface
deletion. We cannot simply get the bridge a BD-VIF is member of via a
YANG subtree filter, as we either get all BDs or only the single bridge
without a bridge id. An xpath filter on the other hand does the trick
and gives us the bridge-id plus the interface as a member in xml.