Don't block on but delete leftover interface NAT
On interface delete the driver used to block the deletion if there was
leftover interface NAT, which will trigger a backoff-retry by NyBase.
This gives another thread the option to delete the leftover NAT if there
is a parallel delete ongoing. But in some cases we just want the NAT
gone, especially when we want to remove a stale interface left on the
router, so we now remove the NAT right away.