Improve and harden router deletion on agent
When delete_router() in neutron is called it gets deleted from the
database. Afterwards all agents are called with a router_deleted()
notification. This puts a router in the processing queue, which then
tries to get router data from neutron - which is no longer there.

To get the router data and do a proper delete (i.e. not let the
scavanger do the work) a copy of the router config is saved in the
neutron cache (which times out after 10 minutes by default). When an
agent is then asking for router config of a deleted router via
get_deleted_router() the result is taken from said cache (keyed by
router id and host).