Don't update already deleted routers
The router processing queue might still have a router in it scheduled
for update (with the router data already present), which might sync us
an already deleted router back on the device. To prevent this we keep a
cache of deleted router ids for about an hour. If a router is
unscheduled and then rescheduled or updated in any other way after the
delete then we remove it from this cache.