Do not delete router atts on router delete
When a router is deleted in the agent it could just got unscheduled from
the agent. While its RouterAtts could be deleted and new ones created on
the next schedule we would still risk a racecondition when the router is
unscheduled. The agent would remove the device config and then request
an unconditional delete of the RouterAtts. If the router has been
rescheduled onto another agent it would loose its RouterAtts. Disabling
the immediate unconditional delete solves this problem. The RouterAtts
will get cleaned up on the next orphan-router-atts-deletion run.