Fix nat double-delete / nat-RPCError
Whenever a floating ip is deleted clean_nat() is called. This method
deletes all no longer needed floating ip nat statements from the device,
if they exist. On every single cleanup, this caused an exception:

RPCError: ...
<bad-element>nat-static-transport-list-with-vrf</bad-element> ...
'severity': 'error', 'tag': 'data-missing'

This is caused by the driver trying do delete the element two times on
each call. The first call succeeds, the second call fails, as the item
is no longer present and we use DELETE. The problem here is the driver
executing that query twice. This can be fixed by not calling delete(),
which always executes a delete on both devices on a pair, but using
_delete(), which is only executed on the device given by context.

Additionally, the delete() of StaticNat is removed, as it serves no
other purpose than the method it overrides (plus it is hiding that the
context passed to it is overridden by execute_on_pair() later).