Replace MutliConfigParser with ConfigParser
MultiConfigParser is deprecated. Sadly the asr1k-address-scopes config
section is in a format where we cannot predict the value of the keys,
thus not allowing us to use regular oslo_config methods for parsing.
Therefore we replace the MultiConfigParser by a rough equivalent, where
we parse each given config file. We return a dict for the first config
file that has a non-empty section with the given name.