Change order of items in YANG dict
Dictionary order in cpython2.7 was alphabetic, in python3.6 it is
insertion order. In theory we should use OrderedDict everywherw where
order matters and not rely on an implementation detail of cpython, but
as python3.7+ makes this implementation detail to a part of the spec I
think we're fine here.

In certain cases the ASR1k netconf-YANG API requires a specific
attribute to appear before all others, e.g. <name> before <service>,
even when these are on the same nesting-level. It looks like they
require keys (tags identifying an item) to come first.