WIP: AZ aware router scheduling
When a router is scheduled with an AZ hint it will only be scheduled
onto an asr1k agent with that given AZ, otherwise it will be scheduled
onto an agent without an AZ set. None and 'nova' (the default AZ) are
both regarded as "no AZ set". Only a single AZ hint is allowed per
router.

A router now also checks if a network has an AZ set on router attach. If
the network has a hint but the router does not then association is
denied.

Agents now properly advertise their given AZ. The ML2 agent does not use
this information at the moment, as for ml2/portbinding AZ restrictions
will be enforced by a driver above the asr1k driver.

WIP because there is some code cleanup to do and it's probably not worth
it to do this for queens, as I won't merge it before we rolled out
ussuri.