Remove Prometheus exporter tests
The Prometheus exporter tests did not work on Travis and do not work
when parallelizing test running. The problem is that the underlying
library prometheus_client has a method start_http_server() which launces
a thread running the http server. There is no easy way to stop this
server, meaning we cannot stop it after a test has run. Therefore we
cannot use setUpClass() to run a single server for each instance (as
with test parallelization each testrunner runs setUpClass once), we
cannot run it once for each test (as we cannot stop the server) and we
don't have an easy way to stop test parallelization without stopping it
for all other tests as well. We could use an extra port for each test
(which doesn't feel very scalable), but we'd still have the running http
servers lingering around afterwards. Therefore I'm going for removing
these tests.