Delete wrong BD membership of BD-VIF on BD update
When adding a BD-VIF to a BridgeDomain (BD) it must not be member of
another bridge, else YANG cannot add it to the new BD. Therefore we need
to add a preflight check for the BridgeDomain model which checks for all
its member if they are in another BD and if they are remove it from
this BD. This is done without calling the preflight check (as we only
want to delete a membership) and without internal validate (as this
could prevent the update from going out, as a single delete request
might look like "nothing changed" to the driver (which is a different
issue, but hasn't fallen on our foots so far)).

Besides this we also need to look into why the device_cleaner is not
properly cleaning bridge memberships, but in any case doing this on
BD update and not only with a cleaner means that calling a sync on a
router (and the regular syncloop can fix this problem).

This problem can happen if for example a delete call fails because YANG
is locking us out again and all retries have been exceeded.