Clean VBInterface even when router id still exists
The cleanup-loop works in 3 steps: It (1) fetches the config objects
from the device, then (2) fetches all relevant objects from the
neutron-server and (3) deletes the objects from the device that it
gathered in (1) but could not find in (2). Object are considered to be
matching if their primary key (in this case BD-VIF ID) matches. Imagine,
a BD-VIF ID was fetched from the device in (1) assigned to VRF A, not
fetched in (2), reassigned after (2) to another VRF. We would then
delete that BD-VIF and break traffic.

In order to prevent this race-condition of deleting a reassigned BD-VIF,
we only delete BD-VIFs if their VRF has been deleted as well. That leads
to a condition, that BD-VIFs could be left orphan, if their VRF is still
existing.

With this PR, we remove the check if a VRF is still existing, so the
BD-VIF will be cleaned up regardless. To prevent the reassignment
race-condition, we fetch the BD-VIF again, lines before it’s deletion
and check if it has been reassigned in the meantime. In this case, do we
still find it assigned to the same VRF seen in (1) than seen now. Hence,
we decrease the time frame in which a race-condition could be happening
to likely milliseconds, subsequently we will only break ports that were
working for milliseconds.