Add policy update to router sync
fwg policies are shared objects, they can be used by multiple routers.
We map them 1:1 to ACL objects on IOS XE and thus need a method to keep
them updated. We could do that by calling a router sync on any single
router that has a reference to that policy or we just call it on an
arbitrary router on every agent. This is what we implement in this
commit.