Create interface RII and zone-membership if external policy present
In order for the ZBF feature to replicate the state table, interfaces
need to be assigned to a redundandancy group and receive a Redundant
Interface Identifier (RII) in order to find the same interface on the
remote side. In case we find an external policy on the `Router`, we
mark all interfaces on the router's interface list with
`has_stateful_firewall`. If that is the case, we instantiate the yang
model with RII, redundancy group membership and in case of the interface
being the gateway interface with an zone membership. As this marking is
done post `Interface` instantiation, we make the `_rest_definition`
attribute a dynamic property, so later changes in the `Interface` object
are represented in its `_rest_definition`.