Make sure to set agent notifier for AGENT_TYPE_L3
For RPC communication with the agent we need to set our own notifier API
instance in agent_notifiers, else we will not get a specific class of
notifications (like router_removed_from_agent()). With the split between
rpc and non-rpc functionality this update was only done in the RPC part.
To allow everyone to read with the L3 agent we now move it into
__init__().