Honor adv. extra routes flag for bgpvpn
By default BGPVPNs advertise all router_routes. BGPVPNRouterAssociation
have a flag that controls this behavior, but we can only implement this
feature on a per-router scope. As this feature was previously turned on
by default and on is also the most common use-case. This also means that
this feature can only be turned off for a router by turning it off on
_all_ of this router's bgpvpn router associations.