AZ aware router scheduling
When a router is scheduled with an AZ hint it will only be scheduled
onto an asr1k agent with that given AZ, otherwise it will be scheduled
onto an agent without an AZ. None and 'nova' (the default AZ) are both
regarded as "no AZ set". Only a single AZ hint is allowed per router.

A router now also checks if a network has an AZ set on router attach. If
the network has a hint but the router does not then association is
denied. This is only done for internal networks. Gateways (i.e. external
networks) are not checked, which allows attached FIPs to move between
AZs. This can be used to move a FIP from one AZ to another.

Agents now properly advertise their given AZ. The ML2 agent does not use
this information at the moment, as for ml2/portbinding AZ restrictions
will be enforced by a driver above the asr1k driver.

The agent AZ check done on scheduling can be disabled with the config
option ignore_invalid_az_hint_for_router. Routers will be scheduled to
any router if no fitting router is present and a warning will be logged.

AZ hints are validated on create to ensure that the referenced AZ
actually exists. The AZ is validated against the set of AZs all
currently running agents have. To disable this check set
ignore_invalid_az_hint_for_router to True. This can be useful when users
want to already add AZ hints to devices for routers still in buildup.
This will also require the router scheduling validation to be disabled.