Allow to overwrite default route via extra routes
If a customer in the past configured a default route it was up to the
yang stack which route would be programmed, the openstack one pointing
to the real northbound gateway or the customers overwrite.

This change will allow explicitly overwritten default routes and removes
the openstack default route from the yang request. Once a default route
was overwritten, outbound connectivity for all networks on that router
will be broken, hence we set a tag indicating to skip the monitoring.
The namespace prober will be extended to make use of this tag.