Throw API error if a networks segment exceeds the BD-VIF limit
The ASR/Catalyst router platform has a limit that does not allow more
than 100 BD-VIFs per bridge-domain, or in our OpenStack setup more than
100 ports in an ASR bound segment. As we cannot work around that, we
will now raise an API exception and tell the user that we cannot assign
that network.

We also take care of this on scheduling in case some of the hardware
routers/agents are above limit while others are not.