Fix incorrect exception when network was not bound to an ASR yet
`get_network_port_count_per_agent()` came back with an empty dict when
the network was not yet bound anywhere. This is an issue as subsequent
checks relied on the count to be present.

In _l3_extension_adapter_ `ensure_bdvif_limit()` evaluated
`all(x >= limit for x in port_count.values()))` where `port_count =
get_network_port_count_per_agent(...)`. This in turn led to the
evaluation of `all([])` which is True. That triggered an exception to
fly when there was no necessity.

If we make sure all ASR agents always show up on return of this query
but with a default of 0, we circumvent this from happening. We also
prevent a KeyError in `ensure_bdvif_limit()` that would occur if a
network was not yet bound on that ASR.