Use replace instead of merge for prefix lists
When a prefix list has multiple members and these members change (e.g.
by removing a subnet from a network) the driver might not be able to
reach the desired config state. Old rules might stay, rules might not be
properly replaced, we might end up with the same rules multiple times
(but with different seqs) on the device. Therefore we now use replace
(aka PUT) instead of merge.