Make stateless NAT mandatory in IOS XE 17.6 and above
The migration for stateless NAT is complicated as it cannot be done
during operations. Hence the game plan is to reboot the passive device
on 17.6, let it be synced with NAT stateless, then failover and do the
dance again.

For this procedure it comes in handy for us to completely ignore any
stateless NAT config and just make it mandatory when 17.6 is running.