Always use no-alias NAT on firmware 17.6 and above
Static NAT creates an implicit arp alias if the NAT IP resides in the ip
nat outside interface's subnet. We also create an explicit arp alias for
every static NAT statement, regardless of subnet affinity. Cisco stated
that on cleanup there might be a condition under which the deletion of
the explicity arp alias and the deletion of the implicit arp alias
(triggered by the deletion of the static arp statement) challange each
other and lead to the arp alias still residing in hardware. This in turn
means the device still replies to arp for that IP.

As a workaround we add each static NAT statement without the implicit
arp alias. We handle the explicit arp alias already, hence now there is
only a single arp alias in the house and the deletion should work as
expected.