Advertise extraroute-atomic extension
We support extraroute-atomic, but don't advertise it properly, which
means it does not appear in the available OpenStack network extensions.
This advertising is needed for clients which want to decide if they can
use that API or not.

Reviewing the whole way we manage supported_extension_aliases shows that
we should probably rework this code and orient ourselves a bit at what
happened in the reference implementation in neutron, but today is not
that day.