Add floating ip ARP cleaning sync loop to agent
The asr1ks in our infrastructure have a relatively high ARP timeout
(default is set to 4 hours). If in this time a fip moves from one router
to another, it changes mac address. This means that routers where this
FIP is not configured on might have a stale entry and will send out IP
packets to the wrong mac in the Ethernet frame. Normally this would be
solved by sending out a gratuitous ARP packet from the router, but the
router does not do this for NAT statements (a long open feature
request). Sending out "fake" ARP requests have either resulted in the
ASR not learning the new mac (IP packet from behind the ASR via network
agent as broadcast) or caused the ACI to either discard the ARP packet
(as it "knows better" what we want) or to relearn the endpoint onto
another device, where we would not want it. Therefore we now try to find
these stale entries on other routers and clean them.

The agent now as an ARP cleaning syncloop. By default every two minutes
(though it is configurable). The agent fetches all fip --> router mac
associations. We expect each fip - if we see it - to be on the mac
address of the gateway interface of our router. Then, for each device on
the agent, we fetch the ARP cache and look for IPs that we know about
and are on the wrong mac. Everything wrong will then be cleared.

Fetching all ARP entries from a device takes about 1.2-4.0s in my
tests, depending on the ARP table size on the device. Selecting only a
subset of fields makes the query go something like 20-40% faster.