Configurable ARP timeout for router interfaces
We can now specify an ARP timeout that gets configured on all router
interfaces, configured separately for external and internal interfaces.
This can be used facilitate faster ARP cleanup of old IP-mac
combinations. The main target of this feature is to get faster FIP reuse
working, where routers are in the same external network. Therefore the
current defaults are 30 minutes (1800s) for external interfaces and
unset for internal interfaces. I'm not currently aware of any problems
inside internal networks, but when the time comes it might be nice to
already have this option waiting.