Ignore non-existant device ARP cache
If the device doesn't give us an ARP cache (e.g. if netconf-yang is
overloaded or disabled) we now don't throw an Attribute error anymore,
but instead log a warning that we couldn't fetch the ARP cache. I opted
for a warning instead of an error as I think YANG problems either are or
should be handled elsewhere, this new could should not be the place to
alert or throw out Sentry errors, as they are (or should be) duplicates.