Use remove instead of delete for bridge members
In some cases we want to delete a bridge member, maybe even as a "side
operation", i.e. triggered by another operation. If the interface has
already been removed (by the cleaning loop, another thread or whatever)
we don't want to fail, but silently ignore this. Therefore "remove"
seems to be the operation which is more fitting.

One could argue that this is the operation we want to use in most cases
in this driver... well, we're getting there.