Don't get subnet CIDR if not present
When instantiating the Router model with router data from Neutron where
the gateway port is unbound we might not have a primary subnet. We
already handle the case that the subnet has no CIDR, but if the subnet
is None, we don't have the .get() method available, resulting in an
AttributeError in _build_pbr_acl(). To fix this we simply check if the
given subnet is None beforehand.