Set GARP flag for FIPs on 17.13
Firmware 17.13 starts to support the garp-interface option for static
NAT configuration. This means that on configuration a gratuitous ARP
will be send out on configuration of the statement. To use this, we need
to be at least on 17.13. The target interface will always be the gateway
interface. Without a gateway interface, the flag will be removed.