Enable native pagination/sorting/filtering
When these three attributes are not set, some part of the filtering and
sorting is done in software in python instead directly via SQL. Upstream
Neutron had these setting for some time and as we're using all the
upstream methods to interact with routers (but do not inherit the
L3Plugin baseclass, as we have our own there) we can set these
attributes as well. It is questionable if this will make a big
performance impact, as routers are not often paginated, but it probably
won't hurt either.