Add API to retrieve router centric view of firewall rules
FWaaS rules are mapped via policies and firewall groups to ports. Yet
when we need to troubelshoot, we must compare it to a virtual router's
configuration, so we'd need easy access to how the rules would look from
a router interface perspective. This commit adds an API endpoint that
delivers that view, removing some cluttering attributes from the result.