Implement cleanup loop for FWAAS objects not tied to a router
Firewall policies can be bound to multiple ports. They are realized by
ACLs, a ClassMap and a ServicePolicy (if the policy is bound to an
external interface). That means, just because a router or a port is
removed from a policy the corresponding objects may still be used from
other routers. Hence it is not removed immediatly but by a cleanup loop,
that checks if anybody on an agent is using that policy and then it will
be removed.