Rework router atts allocation logic
The router_atts (currently mainly the router rd) have previously been
allocated on router create and on every router action (if they did not
exist). Each router has one router_atts entry that has a lifecycle of
being created on router creat and deleted on router delete, so ensuring
it all the time does not make sense from that perspective.

To make sure router_atts are properly allocated on router create they
are now moved into the router precommit hook, where they are allocated
inside the transaction (in comparison to after the transaction, as it
previously was). Also, to clean up the code a bit I replaced the
RouterAttsDb class with ensure_router_atts() inside our db plugin, which
orients itself a little bit more on standard OpenStack conventions.