Implement dynamic NAT pooling for routers
The user can now assign multiple IPs to a router. All but the last IP
are then used as a dynamic NAT pool configured on the router. This
requires all provided IPs to be consecutive. From OpenStack side the
user can add, update and remove the fixed IPs from the router's gateway
interface as they please and the driver will realize or delete the given
NAT pool on device.

This feature collides with the "snat_mode" pool. This mode is used
nowhere in our infrastructure and is most likely also not coming back,
as it requires at least two IPs per router. Therefore all code regarding
this, including the config option for it is removed from the code.