Use redundancy/mapping-id for NAT pools
With NAT pools we can use redundancy/mapping-id on the NAT list, so we
(hopefully) can transport all established NAT sessions by the pool onto
the standby and have smoother failovers with less sessions going down /
needing to be reestablished.

Configuring the session was challenging, as the vrf definition is moved
into the redundancy tag when redundancy is used. As we want to support
configuring this with and without redundancy/mapping-id, we need to
support both ways, which means that we a) need to be able to output
config for each option (easily done with a single if) and b) need to be
able to understand both formats when pulling config from device, which
we do by adding two extra options and, if they're set, we're overriding
the default vrf.